requirements.txt
astroid==2.4.2
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-extra-views==0.13.0
django-import-export==2.4.0
django-model-utils==1.3.1
django-stdimage==5.1.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.4.2
jdcal==1.4.1
jedi==0.17.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
progressbar2==3.51.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python-decouple==3.3
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
tablib==2.0.0
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
